requirements.txt
llama_index
llama-index-embeddings-huggingface
pypdf
python-dotenv
pandas
